+++ PLSS/Models/Coordinate.cs
@@ -29,14 +29,14 @@ namespace PLSS.Models
 
         public Guid CoordinateId { get; private set; }
         public string Northing { get; private set; }
-        public float NorthingDegrees { get; private set; }
-        public float NorthingMinutes { get; private set; }
-        public float NorthingSeconds { get; private set; }
+        public double NorthingDegrees { get; private set; }
+        public double NorthingMinutes { get; private set; }
+        public double NorthingSeconds { get; private set; }
         public string Easting { get; private set; }
-        public float EastingDegrees { get; private set; }
-        public float EastingMinutes { get; private set; }
-        public float EastingSeconds { get; private set; }
-        public float ElipsoidHeight { get; private set; }
+        public double EastingDegrees { get; private set; }
+        public double EastingMinutes { get; private set; }
+        public double EastingSeconds { get; private set; }
+        public double ElipsoidHeight { get; private set; }
         public string Adjustment { get; private set; }
         public string Datum { get; private set; }
